rafactor: 원장 화면 헤더 컴포넌트 파일 분리, component 폴더로 #8
